requirements.txt
fal-client
Pillow
tomli; python_version < "3.11"
